pyproject.toml
[tool.poetry]
name = "fastapi-project"
version = "0.1.0"
description = ""
authors = ["Regina0101 <ritodo1010@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.1"
uvicorn = {extras = ["standard"], version = "^0.30.3"}
requests = "^2.32.3"
sqlalchemy = "^2.0.31"
psycopg2-binary = "^2.9.9"
asyncpg = "^0.29.0"
alembic = "^1.13.2"
databases = "^0.9.0"
greenlet = "^3.0.3"
python-jose = "^3.3.0"
fastapi-jwt-auth = "^0.5.0"
python-multipart = "^0.0.9"
bcrypt = "^4.2.0"
passlib = "^1.7.4"
libgravatar = "^1.0.4"
fastapi-mail = "^1.4.1"
python-dotenv = "^1.0.1"
redis = "^5.0.8"
fastapi-limiter = "^0.1.6"
cloudinary = "^1.41.0"
sphinx = "^8.0.2"
sphinx-autodoc-typehints = "^2.2.3"
pytest = "^8.3.2"
pytest-mock = "^3.14.0"
fakeredis = "^2.23.5"
pytest-asyncio = "^0.23.8"


[tool.poetry.group.dev.dependencies]
sphinx = "^8.0.2"
pytest = "^8.3.2"
pytest-cov = "^5.0.0"


[tool.poetry.group.test.dependencies]
aiosqlite = "^0.20.0"
pytest-asyncio = "^0.23.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
filterwarnings = [
    "ignore::DeprecationWarning:passlib.utils.__init__",
    "ignore::DeprecationWarning:jose.jwt:311"
]
pythonpath = ["."]
